refactor: src 폴더 제거
