+++ calculator_example/parse.py
@@ -1,8 +1,8 @@
 # Yacc example
-from src.ply import yacc
+from ply import yacc
+
 
 # Get the token map from the lexer.  This is required.
-from tokenizer import tokens
 
 def p_expression_plus(p):
     'expression : expression PLUS term'
