+++ src/tokenToConstant.py
